+++ src/Factory.php
@@ -23,11 +23,11 @@ class Factory
     protected $request;
 
     /**
-     * List of templates.
+     * List of parsers.
      *
      * @var array
      */
-    protected $names = [];
+    protected $parsers = [];
 
     /**
      * Construct a new Facile service.
@@ -39,33 +39,11 @@ class Factory
     {
         $this->app = $app;
         $this->request = $request;
-        $this->names = [];
+        $this->parsers = [];
     }
 
     /**
      * Create a new Facile instance.
-     *
-     * <code>
-     *      // Using provided facade for Facile.
-     *
-     *      $users  = User::paginate(30);
-     *      $facile = Facile::make('simple', array(
-     *          'view'   => 'home.index',
-     *          'data'   => array(
-     *              'users' => $users,
-     *          ),
-     *          'status' => 200,
-     *      ));
-     *
-     *      // Alternatively
-     *      $facile = Facile::make('simple')
-     *          ->view('home.index')
-     *          ->with(array(
-     *              'users' => $users,
-     *          ))
-     *          ->status(200)
-     *          ->format('html');
-     * </code>
      *
      * @param  string  $name   Name of template
      * @param  array   $data
@@ -80,18 +58,6 @@ class Factory
 
     /**
      * Create a new Facile instance helper via view.
-     *
-     * <code>
-     *      // Using provided facade for Facile.
-     *
-     *      $users  = User::paginate(30);
-     *      $facile = Facile::view('home.index', array(
-     *              'users' => $users,
-     *          ))
-     *          ->status(200)
-     *          ->template(Orchestra\Facile\Template\Template::class)
-     *          ->format('html');
-     * </code>
      *
      * @param  string  $view
      * @param  array   $data
@@ -108,19 +74,6 @@ class Factory
     /**
      * Create a new Facile instance helper via with.
      *
-     * <code>
-     *      // Using provided facade for Facile.
-     *
-     *      $users  = User::paginate(30);
-     *      $facile = Facile::with(array(
-     *              'users' => $users,
-     *          ))
-     *          ->view('home.index')
-     *          ->status(200)
-     *          ->template(Orchestra\Facile\Template\Template::class)
-     *          ->format('html');
-     * </code>
-     *
      * @param  mixed  $data
      *
      * @return \Orchestra\Facile\Facile
@@ -132,61 +85,89 @@ class Factory
     }
 
     /**
-     * Register a named template.
+     * Register a named parser.
      *
      * @param  string  $name
-     * @param  string|\Orchestra\Facile\Template\Template  $template
+     * @param  string|\Orchestra\Facile\Template\Parser  $parser
      *
      * @return void
      */
-    public function name(string $name, $template): void
+    public function name(string $name, $parser): void
     {
-        if (is_string($template) && (class_exists($template, false) || $this->app->bound($template))) {
-            $template = $this->app->make($template);
+        if (is_string($parser) && (class_exists($parser, false) || $this->app->bound($parser))) {
+            $parser = $this->app->make($parser);
         }
 
-        $this->names[$name] = $template;
+        $this->parsers[$name] = $parser;
     }
 
     /**
      * Get request format.
      *
-     * @param  string  $name
+     * @param  \Orchestra\Facile\Template\Parser  $parser
      *
      * @return string|null
      */
-    public function getRequestFormat($name): ?string
+    protected function getPrefersFor(Template\Parser $parser): ?string
     {
         return $this->request->prefers(
-            $this->getTemplate($name)->getSupportedFormats()
+            $parser->getSupportedFormats()
         );
     }
 
     /**
-     * Get the template.
+     * Resolve parser from factory.
      *
-     * @param  string|\Orchestra\Facile\Template\Template  $name
+     * @param  string  $name
+     * @param  string|null  $format
+     * @param  array  $data
+     * @param  string  $method
      *
-     * @throws \InvalidArgumentException if template is not defined
-     *
-     * @return \Orchestra\Facile\Template\Template
+     * @return mixed
      */
-    public function getTemplate($name): Template\Template
+    public function resolve($name, $format, array $data, $method = 'compose')
     {
-        if ($name instanceof Template\Template) {
+        $parser = $this->parse($name);
+        $format = $format ?? $this->getPrefersFor($parser);
+
+        return $parser->compose($format, $data, $method);
+    }
+
+    /**
+     * Get the parser.
+     *
+     * @param  string|\Orchestra\Facile\Template\Parser  $name
+     *
+     * @throws \InvalidArgumentException if parser is not defined
+     *
+     * @return \Orchestra\Facile\Template\Parser
+     */
+    public function parse($name): Template\Parser
+    {
+        if ($name instanceof Template\Parser) {
             return $name;
-        } elseif (! isset($this->names[$name]) && is_string($name)) {
+        } elseif (! isset($this->parsers[$name]) && is_string($name)) {
             $this->name($name, $name);
         }
 
-        $template = $this->names[$name] ?? $name;
+        $parser = $this->parsers[$name] ?? $name;
 
-        if (! $template instanceof Template\Template) {
+        if (! $parser instanceof Template\Parser) {
             throw new InvalidArgumentException(
-                "Expected \$template to be instanceof Orchestra\Facile\Template\Template."
+                "Expected \$parser to be instanceof Orchestra\Facile\Template\Parser."
             );
         }
 
-        return $template;
+        return $parser;
+    }
+
+    /**
+     * Determine if the application is running unit tests.
+     *
+     * @return bool
+     */
+    protected function runningUnitTests()
+    {
+        return $this->app->runningInConsole() && $this->app->runningUnitTests();
     }
 }