+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This changelog references the relevant changes (bug and security fixes) done to `orchestra/facile`.
 
+## 3.6.1
+
+Released: 2018-05-02
+
+### Added
+
+* Allow certain data to be streamed instead of just rendered.
+
+### Changes
+
+* Revamp `Orchestra\Facile\Template\Template` to `Orchestra\Facile\Template\Parser`.
+
 ## 3.6.0
 
 Released: 2018-02-20
@@ -10,6 +22,18 @@ Released: 2018-02-20
 
 * Update support for Laravel Framework v5.6.
 
+## 3.5.1
+
+Released: 2018-04-02
+
+### Added
+
+* Allow certain data to be streamed instead of just rendered.
+
+### Changes
+
+* Revamp `Orchestra\Facile\Template\Template` to `Orchestra\Facile\Template\Parser`.
+
 ## 3.5.0
 
 Released: 2017-11-10
@@ -17,11 +41,3 @@ Released: 2017-11-10
 ### Changes
 
 * Update support for Laravel Framework v5.5.
-
-## 3.4.0
-
-Released: 2017-01-22
-
-### Changes
-
-* Update support for Laravel Framework v5.4.